pyproject.toml
[build-system]
requires = [
    "setuptools>=61.2",
    "setuptools_scm",
]
build-backend = "setuptools.build_meta"

[project]
name = "glue-wwt"
authors = [{name = "Thomas Robitaille", email = "thomas.robitaille@gmail.com"}]
license = {text = "BSD 3-Clause License"}
description = "Glue WorldWide Telescope plugin"
readme = "README.rst"
urls = {Homepage = "https://github.com/glue-viz/glue-wwt"}
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "glue-core>=1.13.1",
    "echo",
    "astropy",
    "pywwt>=0.24.1",
    "packaging",
    "ipywidgets>=8",
]
dynamic = ["version"]

[tool.setuptools_scm]

[project.entry-points]
"glue.plugins" = {wwt-qt = "glue_wwt.viewer.qt:setup", wwt-jupyter = "glue_wwt.viewer.jupyter:setup"}

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov"
]
qt = [
    "qtpy",
    "glue-qt",
    'PyQt5;python_version>="3"',
    'PyQtWebEngine;python_version>="3"',
]
jupyter = [
    "glue-jupyter",
]

[tool.setuptools]
zip-safe = false
license-files = ["LICENSE"]
include-package-data = false

[tool.setuptools.packages]
find = {namespaces = false}

[tool.setuptools.package-data]
"glue_wwt.viewer" = ["*.html", "*.js", "*.png"]
"glue_wwt.viewer.qt" = ["*.ui"]
"glue_wwt.viewer.qt.tests" = ["data/*.glu"]
